Removed Layout3D code and demos from branch, moved some of the base code back into Interactables and the tile grid into collections. There is other stuff coming that is better.
