--- Assets/MixedRealityToolkit-Examples/Demos/UX/Layout3D.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ac028674247010540b03730ed08f24e7
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
