--- Assets/MixedRealityToolkit-Examples/Demos/UX/Layout3D/Materials.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6d4b292d8d854b04ba88be723556850b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
