--- Assets/MixedRealityToolkit-Examples/Demos/UX/Layout3D/Materials/PanelBackground.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1b6cf9eadb99d72439be60b39c99ba00
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
