--- Assets/MixedRealityToolkit-Examples/Demos/UX/Layout3D/Scenes.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9e3023e0136b7d24bb426d36c5abb4a1
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
