--- Assets/MixedRealityToolkit-Examples/Demos/UX/Layout3D/Scenes/Layout3DGallery.unity.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 62ef83117f5145b43a864fed4024d288
-timeCreated: 1509573767
-licenseType: Pro
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
