--- Assets/MixedRealityToolkit-Examples/Demos/UX/Layout3D/Scripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a04867557a647104a9bf2c6707730430
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
