--- Assets/MixedRealityToolkit-Examples/Demos/UX/Layout3D/Scripts/AutoCycleLayout3DSize.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8a0ea1dbba199714c982b89ec049bc35
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
