--- Assets/MixedRealityToolkit-Examples/Demos/UX/Layout3D/Scripts/Layout3DSizeOutput.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3ae6d13f935c76a4a979dcd6a686b82b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
