--- Assets/MixedRealityToolkit-SDK/Features/UX/Scripts/Layout3D/CycleLayout3DPixelSize.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c7efdd09191714c4888668a62c249f6b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
