--- Assets/MixedRealityToolkit-SDK/Features/UX/Scripts/Layout3D/Layout3DAlign.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: a2d6129faba6ebd44a660a9eee3a4e01
-timeCreated: 1509372348
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
