--- Assets/MixedRealityToolkit-SDK/Features/UX/Scripts/Layout3D/Layout3DBridgeObjects.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 63495ca13e9554f40b566342255b0ba2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
