--- Assets/MixedRealityToolkit-SDK/Inspectors/UX/Layout3D.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 96334f7dfbc4b6a429e1943ca94db6ac
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
